drivers/i2c: implement of i2c in imxrt board
- implement of i2c in imxrt board
