+++ os/board/imxrt1020-evk/include/board.h
@@ -208,8 +208,8 @@
  *   J23 A5 A5/ADC5/SCL  GPIO_AD_B1_00  LPI2C1_SCL
  */
 
-#define GPIO_LPI2C1_SDA   GPIO_LPI2C1_SDA_2	/* GPIO_AD_B1_01 */
-#define GPIO_LPI2C1_SCL   GPIO_LPI2C1_SCL_2	/* GPIO_AD_B1_00 */
+#define GPIO_LPI2C1_SDA   GPIO_LPI2C1_SDA_1	/* GPIO_AD_B1_15_ */
+#define GPIO_LPI2C1_SCL   GPIO_LPI2C1_SCL_1	/* GPIO_AD_B1_14 */
 
 #define GPIO_LPI2C2_SDA   GPIO_LPI2C2_SDA_2	/* GPIO_AD_B1_01 */
 #define GPIO_LPI2C2_SCL   GPIO_LPI2C2_SCL_2	/* GPIO_AD_B1_00 */
